refactor: SRP 원칙을 준수하도록 도메인 분리
