+++ app/src/main/java/woowacourse/shopping/domain/model/CartItemCounter.kt
@@ -2,13 +2,8 @@ package woowacourse.shopping.domain.model
 
 import woowacourse.shopping.view.cartcounter.ChangeCartItemResultState
 
-class CartItemCounter(
-    count: Int = DEFAULT_ITEM_COUNT,
-    isSelected: Boolean = DEFAULT_ITEM_SELECTED,
-) {
+class CartItemCounter(count: Int = DEFAULT_ITEM_COUNT) {
     var itemCount: Int = count
-        private set
-    var isSelected: Boolean = isSelected
         private set
 
     fun increase(){
@@ -28,16 +23,7 @@ class CartItemCounter(
         itemCount = count
     }
 
-    fun selectItem() {
-        isSelected = true
-    }
-
-    fun unSelectItem() {
-        isSelected = false
-    }
-
     companion object {
         const val DEFAULT_ITEM_COUNT = 0
-        const val DEFAULT_ITEM_SELECTED = false
     }
 }
