+++ app/src/main/res/layout/product_quantity_selection.xml
@@ -24,7 +24,7 @@
         app:layout_constraintEnd_toEndOf="parent"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:visibility="@{product.cartItemCounter.isSelected ? View.VISIBLE : View.GONE}"
+        android:visibility="@{product.itemSelector.isSelected ? View.VISIBLE : View.GONE}"
         android:background="@drawable/shape_border_not_line">
 
         <TextView
@@ -66,6 +66,6 @@
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         android:onClick="@{()->onClickCartItemCounter.clickIncrease(product)}"
-        android:visibility="@{product.cartItemCounter.isSelected ? View.GONE : View.VISIBLE}" />
+        android:visibility="@{product.itemSelector.isSelected ? View.GONE : View.VISIBLE}" />
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
